# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "WyAPI-Golang",
    "version": "2.0.0",
    "description": "网易云音乐解析与下载 API"
  },
  "servers": [
    { "url": "/" }
  ],
  "components": {
    "securitySchemes": {
      "ApiToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Token"
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
      "ApiResponse": {
        "type": "object",
        "properties": {
          "code": { "type": "integer" },
          "status": { "type": "integer" },
          "success": { "type": "boolean" },
          "msg": { "type": "string" },
          "message": { "type": "string" },
          "data": { "type": "object" }
        }
      }
    }
  },
  "paths": {
...
    "/api/info": {
      "get": {
        "summary": "API信息",
        "security": [{ "ApiToken": [] }, { "BearerAuth": [] }],
        "responses": {
          "200": { "description": "ok", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiResponse" } } } }
        }
      }
    },
...
    "/search": {
      "post": {
        "summary": "搜索音乐",
        "security": [{ "ApiToken": [] }, { "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "keyword": { "type": "string" },
                  "limit": { "type": "integer" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "ok", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiResponse" } } } }
        }
      }
    },
    "/download": {
      "post": {
        "summary": "下载音乐",
        "security": [{ "ApiToken": [] }, { "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": { "type": "string" },
                  "quality": { "type": "string" },
                  "format": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "ok" }
        }
      }
    }
  }
}